+++ comics/comics/8bittheater.py
@@ -6,16 +6,9 @@ class Meta(MetaBase):
     language = 'en'
     url = 'http://www.nuklearpower.com/'
     start_date = '2001-03-02'
+    end_date = '2010-06-01'
     rights = 'Brian Clevinger'
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2001-03-02'
-    time_zone = -6
-
     def crawl(self, pub_date):
-        page_url = 'http://www.nuklearpower.com/%s/episode/' % (
-            pub_date.strftime('%Y/%m/%d'),)
-        page = self.parse_page(page_url)
-        url = page.src('img[src^="http://www.nuklearpower.com/comics/"]')
-        title = page.alt('img[src^="http://www.nuklearpower.com/comics/"]')
-        return CrawlerImage(url, title)
+        pass # Comic no longer published